+++ scripts/build
@@ -1,7 +1,7 @@
 #!/bin/bash
 echo 'Running eslint ...'
 
-eslint --ignore-path .eslintignore . 2
+node ./node_modules/eslint/bin/eslint.js --ignore-path .eslintignore . 2
 
 if [ $? -ne 0 ]
 then
